Optimize component to segment key mapping
The new mechanism avoids use of a callback function in favor of
specifying a generic type constrained to require the mapping methods
defined. The mapping type is an empty struct, so we can call the mapping
methods on the default value of the struct directly and inlined. This
mechanism improves component random access performance.