Add component-level queries
These allow iterating over components with a single foreach loop.

Benefits:
- Almost as fast as the nested loops in prior checkin
- Much faster than iterating using tuples since component copying is
avoided
- Can both read and write component values

Drawbacks:
- Need to access components via Value1, Value2, etc fields, although a
Values field is also provided to convert to tuple (at a significant
performance cost)
- Possible GC impact due to reference type enumerator, although this
appears to be minimal and could be pooled if it becomes an issue

Overall, this looks like a good default option that's relatively
straightforward and gives good performance.